+++ pyengine/Widgets/__init__.py
@@ -7,3 +7,4 @@ from pyengine.Widgets.ProgressBar import ProgressBar
 from pyengine.Widgets.Console import Console
 from pyengine.Widgets.MultilineLabel import MultilineLabel
 from pyengine.Widgets.AnimatedImage import AnimatedImage
+from pyengine.Widgets.Selector import Selector