+++ Changelog.md
@@ -14,6 +14,7 @@
     - MultilineLabel : Create Label widget for multiline
     - Label : Replace \n to void. Label doesn't support break lines
     - AnimatedImage : Create Image widget which support multiple images
+    - Selector : Create selector widget
     - Others : Rename Maths to Others
     - Others : Add wrap_text and get_images_from_gif functions
     - Unit Tests : Fix Vec2 and Color tests